+++ src/Proxy/CurvePolygonProxy.php
@@ -325,6 +325,15 @@ class CurvePolygonProxy extends CurvePolygon implements ProxyInterface
         return $this->proxyGeometry->withoutM();
     }
 
+    public function withRoundedCoordinates(int $precision): \Brick\Geo\Geometry
+    {
+        if ($this->proxyGeometry === null) {
+            $this->load();
+        }
+
+        return $this->proxyGeometry->withRoundedCoordinates($precision);
+    }
+
     public function swapXy(): \Brick\Geo\Geometry
     {
         if ($this->proxyGeometry === null) {