+++ src/Proxy/MultiLineStringProxy.php
@@ -307,6 +307,15 @@ class MultiLineStringProxy extends MultiLineString implements ProxyInterface
         return $this->proxyGeometry->withoutM();
     }
 
+    public function withRoundedCoordinates(int $precision): \Brick\Geo\Geometry
+    {
+        if ($this->proxyGeometry === null) {
+            $this->load();
+        }
+
+        return $this->proxyGeometry->withRoundedCoordinates($precision);
+    }
+
     public function swapXy(): \Brick\Geo\Geometry
     {
         if ($this->proxyGeometry === null) {