+++ src/Proxy/PointProxy.php
@@ -280,6 +280,15 @@ class PointProxy extends Point implements ProxyInterface
         return $this->proxyGeometry->withoutM();
     }
 
+    public function withRoundedCoordinates(int $precision): \Brick\Geo\Geometry
+    {
+        if ($this->proxyGeometry === null) {
+            $this->load();
+        }
+
+        return $this->proxyGeometry->withRoundedCoordinates($precision);
+    }
+
     public function swapXy(): \Brick\Geo\Geometry
     {
         if ($this->proxyGeometry === null) {